requirements.txt
discord.py==1.3.2
psycopg2-binary==2.8.4